Separate file for SigningInput
