+++ src/Bitcoin/Transaction.h
@@ -1,4 +1,4 @@
-// Copyright © 2017-2020 Trust Wallet.
+// Copyright © 2017-2021 Trust Wallet.
 //
 // This file is part of Trust. The full Trust copyright notice, including
 // terms governing use, modification, and redistribution, is contained in the
@@ -19,7 +19,6 @@
 #include "../proto/Bitcoin.pb.h"
 
 #include <vector>
-#include <optional>
 
 namespace TW::Bitcoin {
 
@@ -108,50 +107,6 @@ private:
     /// Generates the signature hash for for scripts other than witness scripts.
     Data getSignatureHashBase(const Script& scriptCode, size_t index,
                               enum TWBitcoinSigHashType hashType) const;
-};
-
-// TODO move to separate file
-class SigningInput {
-public:
-    // Hash type to use when signing
-    enum TWBitcoinSigHashType hashType = TWBitcoinSigHashTypeAll;
-
-    // Amount to send.  Transaction created will have this amount in its output, 
-    // except when use_max_amount is set, in that case this amount is not relevant, maximum possible amount will be used (max avail less fee).
-    // If amount is equal or more than the available amount, also max amount will be used.
-    Amount amount = 0;
-
-    // Transaction fee per byte
-    Amount byteFee = 0;
-
-    // Recipient's address
-    std::string toAddress;
-
-    // Change address
-    std::string changeAddress;
-
-    // Available private keys
-    std::vector<PrivateKey> privateKeys;
-
-    // Available redeem scripts indexed by script hash
-    std::map<std::string, Script> scripts;
-
-    // Available unspent transaction outputs
-    UTXOs utxos;
-
-    // If sending max amount
-    bool useMaxAmount = false;
-
-    // Coin type (forks)
-    TWCoinType coinType = TWCoinTypeBitcoin;
-
-    // Optional transaction plan
-    std::optional<TransactionPlan> plan;
-
-public:
-    SigningInput() = default;
-
-    SigningInput(const Proto::SigningInput& input);
 };
 
 } // namespace TW::Bitcoin
