+++ src/Bitcoin/TransactionBuilder.h
@@ -1,4 +1,4 @@
-// Copyright © 2017-2020 Trust Wallet.
+// Copyright © 2017-2021 Trust Wallet.
 //
 // This file is part of Trust. The full Trust copyright notice, including
 // terms governing use, modification, and redistribution, is contained in the
@@ -6,6 +6,7 @@
 
 #pragma once
 
+#include "SigningInput.h"
 #include "Transaction.h"
 #include "TransactionPlan.h"
 #include "UnspentSelector.h"
