+++ tests/Bitcoin/TxComparisonHelper.h
@@ -1,4 +1,4 @@
-// Copyright © 2017-2020 Trust Wallet.
+// Copyright © 2017-2021 Trust Wallet.
 //
 // This file is part of Trust. The full Trust copyright notice, including
 // terms governing use, modification, and redistribution, is contained in the
@@ -7,6 +7,7 @@
 #pragma once
 
 #include "Bitcoin/Amount.h"
+#include "Bitcoin/SigningInput.h"
 #include "Bitcoin/Transaction.h"
 #include "Bitcoin/TransactionPlan.h"
 #include <TrustWalletCore/TWCoinType.h>
